refactor(whatsrust): extract action wrappers
